sonar (#26)
Merging despite a couple items:

- coverage tracking isn't working yet (it will be something simple) 
- there are a ton of these in the sonar log and maybe it's worth fixing.

```
The configuration :service:detachedConfiguration32 is both resolvable and consumable. This is considered a legacy configuration and it will eventually only be possible to be one of these.
The configuration :service:detachedConfiguration32 is both consumable and declarable. This combination is incorrect, only one of these flags should be set.
```